+++ src/components/HomePage.js
@@ -1,17 +1,22 @@
 import React from "react";
 import Articles from "./Articles";
 import TagsList from "./TagsList";
+import AuthHeader from "./AuthHeader";
+import NonAuthHeader from "./NonAuthHeader";
 
 function HomePage() {
   return (
-    <div className="flex">
-      <div className="articleDiv">
-        <Articles />
+    <>
+      <NonAuthHeader />
+      <div className="flexA-T container">
+        <div className="articleDiv">
+          <Articles />
+        </div>
+        <div className="tagsDiv">
+          <TagsList />
+        </div>
       </div>
-      <div className="tagsDiv">
-        <TagsList />
-      </div>
-    </div>
+    </>
   );
 }
