+++ src/components/NonAuthHeader.js
@@ -1,27 +1,9 @@
 import React from "react";
 import { NavLink } from "react-router-dom";
 
-export default function AuthHeader() {
+function NonAuthHeader() {
   return (
-    <div className="header">
-      <header className="flex">
-        <h1 className="headerTitle">CONDUIT</h1>
-        <nav className="navLinks flex">
-          <NavLink to="/" className="inactiveNav" activeClassName="activeNav">
-            Home
-          </NavLink>
-          <NavLink to="#" className="inactiveNav" activeClassName="activeNav">
-            New
-          </NavLink>
-        </nav>
-      </header>
-    </div>
-  );
-}
-
-export default function NonAuthHeader() {
-  return (
-    <div className="header">
+    <div className="headerDiv">
       <header className="flex">
         <h1 className="headerTitle">CONDUIT</h1>
         <nav className="navLinks flex">
@@ -44,3 +26,5 @@ export default function NonAuthHeader() {
     </div>
   );
 }
+
+export default NonAuthHeader;
